terraform/app/requirements.txt
fastapi
uvicorn
google-cloud-pubsub
faker
unidecode
